[unittest] Created a shared library for factories to reduce memory during compilation
