+++ unittest/factory/action.hpp
@@ -6,16 +6,13 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
+#ifndef CROCODDYL_ACTION_FACTORY_HPP_
+#define CROCODDYL_ACTION_FACTORY_HPP_
+
 #include <iterator>
 
 #include "crocoddyl/core/action-base.hpp"
-#include "crocoddyl/core/actions/unicycle.hpp"
-#include "crocoddyl/core/actions/lqr.hpp"
 #include "crocoddyl/core/numdiff/action.hpp"
-#include "crocoddyl/core/utils/exception.hpp"
-
-#ifndef CROCODDYL_ACTION_FACTORY_HPP_
-#define CROCODDYL_ACTION_FACTORY_HPP_
 
 namespace crocoddyl {
 namespace unittest {
@@ -32,58 +29,17 @@ struct ActionModelTypes {
   }
   static const std::vector<Type> all;
 };
-const std::vector<ActionModelTypes::Type> ActionModelTypes::all(ActionModelTypes::init_all());
 
-std::ostream& operator<<(std::ostream& os, ActionModelTypes::Type type) {
-  switch (type) {
-    case ActionModelTypes::ActionModelUnicycle:
-      os << "ActionModelUnicycle";
-      break;
-    case ActionModelTypes::ActionModelLQRDriftFree:
-      os << "ActionModelLQRDriftFree";
-      break;
-    case ActionModelTypes::ActionModelLQR:
-      os << "ActionModelLQR";
-      break;
-    case ActionModelTypes::NbActionModelTypes:
-      os << "NbActionModelTypes";
-      break;
-    default:
-      break;
-  }
-  return os;
-}
+std::ostream& operator<<(std::ostream& os, ActionModelTypes::Type type);
 
 class ActionModelFactory {
  public:
-  ActionModelFactory(ActionModelTypes::Type type) {
-    switch (type) {
-      case ActionModelTypes::ActionModelUnicycle:
-        nx_ = 3;
-        nu_ = 2;
-        action_ = boost::make_shared<crocoddyl::ActionModelUnicycle>();
-        break;
-      case ActionModelTypes::ActionModelLQRDriftFree:
-        nx_ = 80;
-        nu_ = 40;
-        action_ = boost::make_shared<crocoddyl::ActionModelLQR>(nx_, nu_, true);
-        break;
-      case ActionModelTypes::ActionModelLQR:
-        nx_ = 80;
-        nu_ = 40;
-        action_ = boost::make_shared<crocoddyl::ActionModelLQR>(nx_, nu_, false);
-        break;
-      default:
-        throw_pretty(__FILE__ ": Wrong ActionModelTypes::Type given");
-        break;
-    }
-  }
+  ActionModelFactory(ActionModelTypes::Type type);
+  ~ActionModelFactory();
 
-  ~ActionModelFactory() {}
+  boost::shared_ptr<crocoddyl::ActionModelAbstract> create();
 
-  boost::shared_ptr<crocoddyl::ActionModelAbstract> create() { return action_; }
-
-  const std::size_t& get_nx() { return nx_; }
+  const std::size_t& get_nx();
 
  private:
   std::size_t nx_;
