+++ unittest/factory/actuation.hpp
@@ -6,13 +6,11 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
-#include "state.hpp"
-#include "crocoddyl/core/actuation-base.hpp"
-#include "crocoddyl/multibody/actuations/full.hpp"
-#include "crocoddyl/multibody/actuations/floating-base.hpp"
-
 #ifndef CROCODDYL_ACTUATION_FACTORY_HPP_
 #define CROCODDYL_ACTUATION_FACTORY_HPP_
+
+#include "state.hpp"
+#include "crocoddyl/core/actuation-base.hpp"
 
 namespace crocoddyl {
 namespace unittest {
@@ -29,51 +27,16 @@ struct ActuationModelTypes {
   }
   static const std::vector<Type> all;
 };
-const std::vector<ActuationModelTypes::Type> ActuationModelTypes::all(ActuationModelTypes::init_all());
 
-std::ostream& operator<<(std::ostream& os, ActuationModelTypes::Type type) {
-  switch (type) {
-    case ActuationModelTypes::ActuationModelFull:
-      os << "ActuationModelFull";
-      break;
-    case ActuationModelTypes::ActuationModelFloatingBase:
-      os << "ActuationModelFloatingBase";
-      break;
-    case ActuationModelTypes::NbActuationModelTypes:
-      os << "NbActuationModelTypes";
-      break;
-    default:
-      break;
-  }
-  return os;
-}
+std::ostream& operator<<(std::ostream& os, ActuationModelTypes::Type type);
 
 class ActuationModelFactory {
  public:
-  explicit ActuationModelFactory() {}
-  ~ActuationModelFactory() {}
+  explicit ActuationModelFactory();
+  ~ActuationModelFactory();
 
   boost::shared_ptr<crocoddyl::ActuationModelAbstract> create(ActuationModelTypes::Type actuation_type,
-                                                              StateModelTypes::Type state_type) {
-    boost::shared_ptr<crocoddyl::ActuationModelAbstract> actuation;
-    StateModelFactory factory;
-    boost::shared_ptr<crocoddyl::StateAbstract> state = factory.create(state_type);
-    boost::shared_ptr<crocoddyl::StateMultibody> state_multibody;
-    switch (actuation_type) {
-      case ActuationModelTypes::ActuationModelFull:
-        state_multibody = boost::static_pointer_cast<crocoddyl::StateMultibody>(state);
-        actuation = boost::make_shared<crocoddyl::ActuationModelFull>(state_multibody);
-        break;
-      case ActuationModelTypes::ActuationModelFloatingBase:
-        state_multibody = boost::static_pointer_cast<crocoddyl::StateMultibody>(state);
-        actuation = boost::make_shared<crocoddyl::ActuationModelFloatingBase>(state_multibody);
-        break;
-      default:
-        throw_pretty(__FILE__ ":\n Construct wrong ActuationModelTypes::Type");
-        break;
-    }
-    return actuation;
-  }
+                                                              StateModelTypes::Type state_type);
 };
 
 }  // namespace unittest
